refactored files and folders into the flaskr #40
