+++ flaskr/static/CSS/login.css
@@ -1,0 +1,67 @@
+h1 {
+  text-align: center;
+  background: white;
+  font-size: 25px;
+  font-family: Arial;
+  color: #8e0000;
+  border-style: solid;
+  border-color: #8e0000;
+  padding: 30px;
+}
+
+body {
+    background-image: url(https://www.jhharchitects.com//images/work/McCool-Pan_1200x800.jpg);
+    height: 100%
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+}
+
+.login {
+    height: 400px;
+    width: 400px;
+    margin: 200px auto 0;
+    border-style: solid;
+    border-color: #8e0000;
+    text-align: center;
+    font-family: Arial;
+    background: white;
+}
+
+#email {
+  background-color: #f1f1f1;
+  width: 90%;
+  padding: 15px;
+  margin: 8px 0;
+  border: 1px solid #8e0000;
+  border-radius: 4px;
+  box-sizing: border-box;
+  font-family: Arial;
+}
+
+#password {
+  background-color: #f1f1f1;
+  width: 90%;
+  padding: 15px;
+  margin: 8px 0;
+  border: 1px solid #8e0000;
+  border-radius: 4px;
+  box-sizing: border-box;
+  font-family: Arial;
+}
+
+#submit {
+padding: 10px;
+  margin: 8px 0;
+  border: none;
+  border-radius: 4px;
+  width: 90%;
+  background-color: #8e0000;
+  color: white;
+  font-family: Arial;
+}
+
+#submit:hover {
+  background-color: #7e0000;
+  color: #f4f4f4;
+}
