+++ flaskr/static/CSS/message.css
@@ -1,0 +1,18 @@
+#body {
+  text-align: center;
+  background: white;
+  border-style: solid;
+  border-color: #8e0000;
+  padding: 30px;
+}
+
+#backtoinbox {
+  padding: 5px;
+  margin: 0px 0;
+  border: none;
+  border-radius: 4px;
+  background-color: #8e0000;
+  font-size: px;
+  font-family: Arial;
+  color: white;
+}
