+++ flaskr/templates/composition.html
@@ -1,0 +1,32 @@
+{% extends "base.html" %}
+{% block title %}Composition Page{% endblock %}
+{% block content %}
+<!-- TODO: TODO: Add nice layout for the composition page -->
+<head>
+  <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='CSS/composition.css') }}">
+</head>
+  <h1>Send a message with Bully Mail!</h1>
+<body>
+<p><button id="backtoinbox" onclick="window.location.href='/inbox'">Back to inbox</button></p>
+<form action="#" method="post" enctype=multipart/form-data>
+  <!-- Using the input and text area to get the user's input via the browser -->
+  <p><input type="text" placeholder="To: " name="receiver" id="receiver" required/></p>
+  <p><input type="text" placeholder="Subject: " name="subject" id="subject"/></p>
+
+  <!-- Using the Tiny MCE UI to improve the email composition -->
+  <script src="https://cdn.tiny.cloud/1/vr2x3cxjja3hokctzkzvlvxa1ex2uerrs733r5889trcniq3/tinymce/5/tinymce.min.js" referrerpolicy="origin"></script>
+
+  <script>
+ tinymce.init({
+  selector: 'textarea',  // change this value according to your HTML
+  menubar: 'edit view format'
+});
+</script>
+
+  <p><textarea name=emailbody rows="35" cols="100"></textarea></p>
+  <label for="infile">Attach a file:</label>
+  <input type="file" name="infile" id="infile" multiple="">
+  <p><input type="submit" value="Send" name="Send message" id="send"/></p>
+</form>
+</body>
+{% endblock %}
