--- templates/base.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    {% block head %}
-    {% endblock %}
-    <title>{% block title %}{% endblock %}</title>
-  </head>
-  <body>
-    {% block content %}
-    {% endblock %}
-  </body>
-</html>
