--- templates/inbox.html
@@ -1,29 +1,0 @@
-{% extends "base.html" %}
-{% block head %}
-<script src="{{ url_for('static', filename='js/inbox.js') }}"></script>
-{% endblock %}
-{% block title %}Inbox Page{% endblock %}
-{% block content %}
-<!-- TODO: Add nice layout for the inbox page -->
-<head>
-  <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='CSS/inbox.css') }}">
-</head>
-<h1>Inbox</h1>
-{% if error %}
-<p><strong>Error: </strong>{{ error }}</p>
-{% endif %}
-<p><button id="tologout" onclick="window.location.href='/logout'">Logout</button></p>
-<p><button id="tocompose" onclick="window.location.href='/compose/'">Compose</button></p>
-<form>
-  <input type="text" placeholder="Search mail" id="search" />
-  <button class="button">Search</button>
-  <span></span>
-</form>
-
-<!-- Inbox messages -->
-<main>
-</main>
-<p><a href="#" id="loadMoreMessages">Load more messages</a></p>
-<p id="loadAlert"></p>
-
-{% endblock %}
