--- templates/login.html
@@ -1,24 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Login Page{% endblock %}
-{% block content %}
-<!-- TODO: Add nice layout for the login page -->
-<head>
-  <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='CSS/login.css') }}">
-</head>
-<h1>Welcome to Bully Mail!</h1>
-<body>
-<div class="login">
-<h2 id="l">Login</h2>
-{% if error %}
-<p><strong>Error: </strong>{{ error }}</p>
-{% endif %}
-<form action="#" method="post">
-  <!-- Even though the input check for "valid email address",
-    the email address needs to be validated on the server side -->
-  <p><input type="email" placeholder="Enter email" name="email" id="email" required/></p>
-  <p><input type="password" placeholder="Enter password" name="password" id="password" required/></p>
-  <p><input type="submit" name="submit" value="Submit" id="submit" /></p>
-</form>
-  </div>
-</body>
-{% endblock %}
