--- templates/message.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% block head %}
-<script src="{{ url_for('static', filename='js/message.js' )}}"></script>
-{% endblock %}
-{% block title %}{{ subject }}{% endblock %}
-{% block content %}
-<p><a href="{{ url_for('inbox') }}">Go Back</a></p>
-<p><a href="{{ url_for('forward', uid=uid) }}">Forward</a></p>
-<p id="infoAlert">Loading message...</p>
-<main>
-  <h1 id="subject" uid="{{ uid }}"></h1>
-  <p id="from">From: </p>
-  <p id="to">To: </p>
-  <p id="date">Date: </p>
-  <p id="body"></p>
-</main>
-{% endblock %}
